requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
blinker==1.5
Brotli==1.0.9
browser-cookie3==0.16.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
Django==4.1.1
django-tinymce==3.5.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
jaraco.classes==3.2.2
jeepney==0.8.0
Jinja2==3.1.2
kaitaistruct==0.10
keyring==23.9.3
lz4==4.0.2
MarkupSafe==2.1.1
more-itertools==8.14.0
mysqlclient==2.1.1
outcome==1.2.0
pbkdf2==1.3
Pillow==9.2.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pydivert==2.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-decouple==3.6
python-dotenv==0.21.0
pywin32-ctypes==0.2.0
requests==2.28.1
SecretStorage==3.3.3
selenium==4.4.3
selenium-wire==4.6.5
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stripe==4.1.0
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.2
urllib3==1.26.12
webdriver-manager==3.8.3
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.8.1
zstandard==0.18.0
